feat(project): changed package name to com.foodbank.foodbank
