+++ app/src/androidTest/java/com/foodbank/foodbank/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.emilio.foodbank
+package com.foodbank.foodbank
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
