+++ app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.emilio.foodbank">
+    package="com.foodbank.foodbank">
 
     <application
         android:allowBackup="true"
@@ -14,7 +14,7 @@
         android:theme="@style/Theme.FoodBank"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name="com.foodbank.foodbank.MainActivity"
             android:exported="true"
             android:label="@string/app_name">
             <intent-filter>
