+++ app/src/main/java/com/foodbank/foodbank/ui/dashboard/DashboardViewModel.kt
@@ -1,13 +1,13 @@
-package com.emilio.foodbank.ui.home
+package com.foodbank.foodbank.ui.dashboard
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class HomeViewModel : ViewModel() {
+class DashboardViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is home Fragment"
+        value = "This is dashboard Fragment"
     }
     val text: LiveData<String> = _text
 }
