+++ app/src/test/java/com/foodbank/foodbank/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.emilio.foodbank
+package com.foodbank.foodbank
 
 import org.junit.Test
